pyproject.toml
[project]
name = "detect-hatespeech"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "google-generativeai>=0.8.5",
    "python-dotenv>=1.1.1",
]
